Preserve the original swagger during download
It looks like the CLI tool is re-writing the swagger JSON file.  So, folks see a difference in the behavior (esp the ordering of parameters).  We now have quite a few partners reporting this issue.
This is just a temporary fix to preserve the original swagger until we have a proper fix.