Smart Global Governance Connector (#669)
* create folder and apiDefinitionSwagger

* Add files via upload

Initial delivery

* Delete icon.png

On demand of Microsoft. Not needed for open sourcing.

* Delete settings.json

On demand of Microsoft. Not needed for open sourcing.

* Add files via upload

Title of the connector changed to "Smart Global Governance".
Description expanded to highlight the functionality and value that the connector offers.
Contact information changed to "Smart Global Governance Support Team".

* Add files via upload

Added new responses according to HTTP codes.