Rebranding of Projectplace to ProjectPlace
No actual changes - just the rebranding of Planview Projectplace to
camel case ProjectPlace.

Also took the time to update some of the swagger definition file to
refer to "activities" rather than "planlets". The word "planlet" is
really internal to Projectplace - while the word "activity" is
understandable by an end user.