Update to Tulip Connector to add ability to handle more Tulip data types. (#2797)
* Create readme.md

* Create readme.md

* Create apiDefinition.swagger.json

* Create apiProperties.json

* Update readme.md

* Create script.csx

* Create settings.json

* Tulip R1

* New Icon File

* Updating icon

* Updating icon

* Add files via upload

* New Icon

* Add files via upload

* Delete settings.json

Not required in public repsository.

* Delete icon.png

Not required in open source repository

* Fixing typos

* Delete custom-connectors/Tulip directory

Removed because it was in the wrong location

* Delete icon.png

Saved in wrong location

* Updated data type mapping 

Now covers all data types that are available in Tulip Tables except for linked records.

* Added filed mappings and tips for use

Added a table to explain the field / variable type mappings between Tulip tables and Power Automate. Explained the requirements for each and tips for use.