Add files via upload (#2856)
* Add files via upload

Updated IA-Connect Microsoft Office for release 9.4

* Update apiDefinition.swagger.json

Didn't do anything. Just trying to get the checks to re-run

* Update apiDefinition.swagger.json

Returned the output schema for 200 responses back to the original {} for existing actions (so the schema of existing actions doesn't change)

* Update apiDefinition.swagger.json

Reverted one more output schema back to the way it was set in 9.1. This ensures that the output schema doesn't change between versions