array data for DataSchema
Got the swagger right (was really the code was had checked in before with just a little cleanup.

The custom connector is now failing due to internal server error so we need to find a way to use the test page in powerapps online, despite the fact that it doesn't really handle array data very well. Possibly specifying the raw body data might be a workaround.